chore: moved things into a more correct location, not sure about the query parameter invalidation
